03_structured/pyproject.toml
[project]
name = "03-structured"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-core>=0.12.44",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-readers-file>=0.4.9",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.1",
    "typing>=3.10.0.0",
]
